collapsed all card converters into convert-card.js
