--- packages/sites/src/layout/convert-follow-card.ts
@@ -1,6 +1,0 @@
-import { ICard, ITemplatizedCard } from "./types";
-
-export const convertFollowCard = function convertFollowCard (card: ICard): ITemplatizedCard {
-  card.component.settings.initiativeId = '{{initiative.item.id}}';
-  return { card, assets: [] };
-};
